refactor: extract SessionLibraryPresenters.swift from SessionLibraryController.swift
Extracts 4 status-presentation types (DisplayedTranscriptCopyResult enum,
DisplayedTranscriptCopyStatusPresenter enum, TranscriptSaveStatusPresenter
enum, SessionDeletionStatusPresenter enum) + SessionLibraryStatusPresenter
@MainActor class from SessionLibraryController.swift into a new file.

Byte-identical move; SHA-256 5ccb844d0eb90df826fd05ee3f0924e6ec0f3f1a4dee8b59da6bea8479add001
verified against origin/main:SessionLibraryController.swift lines 4-79.

Reduces SessionLibraryController.swift from 325 to 250 lines.

Applied memory rule "watch for annotations on line immediately preceding
deleted blocks": verified @MainActor at line 80 (for SessionLibraryController
class) is preserved before deleting the presenter block 4-79.

Tests: 667 (unchanged).
